Refactor query generators into a separate file
